x/tools/cmd/guru: Add flag -allowerrors
Allow the caller to perform the program analysis even
if some of the packages in the scope contained I/O, parser or type
errors.

For that it sets the flag loader.Config.AllowErrors. The Loader will
still print the errors to stderr, but continue and ignore those
packages with errors.